Fixed test suite compilation on GNUstep and included TestPersistency.m in the GNUmakefile
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@8043 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Tests/TestCompositeLayout.m
@@ -190,7 +190,7 @@ DEALLOC(DESTROY(itemFactory); DESTROY(item))
 }
 
 - (ETLayoutItem *) baseItem: (ETLayoutItemGroup *)baseItem 
-                itemAtIndex: (int)index
+                itemAtIndex: (NSUInteger)index
                 inItemGroup: (ETLayoutItemGroup *)itemGroup
 {
 	return [itemFactory itemWithRepresentedObject: [[self modelContent] objectAtIndex: index]];
